microservices/msgpack-python/requirements.txt
fastapi==0.135.1
uvicorn==0.36.0
psycopg2-binary==2.9.9
pydantic==2.6.3
python-dotenv==1.0.1
msgpack==1.0.7
requests==2.32.3
